# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hearth — family AI workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface ProfileInput {
  /** Free-text description of user's role and use cases */
  rawText: string;
}

export interface Profile {
  id: string;
  userId: string;
  rawText: string;
  role: string;
  useCases: string[];
  tone: string;
  notionConnected?: boolean;
  /** @nullable */
  notionAccessToken?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface Note {
  id: string;
  userId: string;
  title: string;
  content: string;
  notionSynced: boolean;
  /** @nullable */
  notionPageId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ItemType = typeof ItemType[keyof typeof ItemType];


export const ItemType = {
  list: 'list',
  table: 'table',
  calculation: 'calculation',
  business_record: 'business_record',
  note: 'note',
} as const;

export type ItemStructuredData = { [key: string]: unknown };

export interface Item {
  id: string;
  userId: string;
  rawText: string;
  type: ItemType;
  title: string;
  structuredData: ItemStructuredData;
  /** @nullable */
  total: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Dashboard {
  totalNotes: number;
  totalItems: number;
  itemsByType: TypeCount[];
  recentNotes: Note[];
  recentItems: Item[];
  /** Total value of all auction bids */
  auctionTotal: number;
}

export interface NoteInput {
  title: string;
  content?: string;
}

export interface NoteUpdate {
  title?: string;
  content?: string;
}

export interface ItemInput {
  rawText: string;
}

export type ItemUpdateStructuredData = { [key: string]: unknown };

export interface ItemUpdate {
  title?: string;
  rawText?: string;
  structuredData?: ItemUpdateStructuredData;
}

export interface ItemsSummary {
  total: number;
  byType: TypeCount[];
}

export interface AuctionEntryInput {
  item: string;
  winningBid: number;
  buyer: string;
  date: string;
}

export interface AuctionEntryUpdate {
  item?: string;
  winningBid?: number;
  buyer?: string;
  date?: string;
}

export interface AuctionEntry {
  id: string;
  userId: string;
  item: string;
  winningBid: number;
  buyer: string;
  date: string;
  createdAt: string;
}

export interface ShippingInput {
  origin: string;
  destination: string;
  weightKg: number;
  description?: string;
}

export interface ShippingCalculation {
  id: string;
  userId: string;
  origin: string;
  destination: string;
  weightKg: number;
  /** @nullable */
  description?: string | null;
  estimatedCost: number;
  currency: string;
  createdAt: string;
}

export interface LinkInput {
  url: string;
  label?: string;
}

export interface BusinessLink {
  id: string;
  userId: string;
  url: string;
  label: string;
  aiSummary: string;
  createdAt: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AIChatInput {
  message: string;
  history?: ChatMessage[];
}

export interface AIChatResponse {
  message: string;
  provider: string;
}

export interface AIProcessInput {
  text: string;
}

export type AIProcessResponseType = typeof AIProcessResponseType[keyof typeof AIProcessResponseType];


export const AIProcessResponseType = {
  list: 'list',
  table: 'table',
  calculation: 'calculation',
  business_record: 'business_record',
  note: 'note',
} as const;

export type AIProcessResponseStructuredData = { [key: string]: unknown };

export interface AIProcessResponse {
  type: AIProcessResponseType;
  title: string;
  structuredData: AIProcessResponseStructuredData;
  /** @nullable */
  total: number | null;
  provider: string;
}

export type ListNotesParams = {
search?: string;
notionSynced?: boolean;
};

export type ListItemsParams = {
type?: ListItemsType;
search?: string;
};

export type ListItemsType = typeof ListItemsType[keyof typeof ListItemsType];


export const ListItemsType = {
  list: 'list',
  table: 'table',
  calculation: 'calculation',
  business_record: 'business_record',
  note: 'note',
} as const;