# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hearth — family AI workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PROFILE
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: notionSynced
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notes/{id}/sync-notion:
    post:
      operationId: syncNoteToNotion
      tags: [notes]
      summary: Sync note to Notion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Notion not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ITEMS (Universal Input)
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List structured items
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [list, table, calculation, business_record, note]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"

    post:
      operationId: createItem
      tags: [items]
      summary: Create an item from natural language input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
...
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"

    patch:
      operationId: updateItem
      tags: [items]
      summary: Update an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"

    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # AUCTION
...
  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List business links
      responses:
        "200":
          description: Links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessLink"

    post:
      operationId: createLink
      tags: [links]
      summary: Save a business link with AI-generated summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessLink"
  /links/{id}:
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a business link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # AI
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # AUTH
    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    # PROFILE
...
    NoteInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    Note:
      type: object
      required: [id, userId, title, content, notionSynced, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        notionSynced:
          type: boolean
        notionPageId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ITEMS
    ItemInput:
      type: object
      required: [rawText]
      properties:
        rawText:
          type: string
    ItemUpdate:
      type: object
      properties:
        title:
          type: string
        rawText:
          type: string
        structuredData:
          type: object
          additionalProperties: true
    Item:
      type: object
      required: [id, userId, rawText, type, title, structuredData, total, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        rawText:
          type: string
        type:
          type: string
          enum: [list, table, calculation, business_record, note]
        title:
          type: string
        structuredData:
          type: object
          additionalProperties: true
        total:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LinkInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        label:
          type: string
    BusinessLink:
      type: object
      required: [id, userId, url, label, aiSummary, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        url:
          type: string
        label:
          type: string
        aiSummary:
          type: string
        createdAt:
          type: string
          format: date-time

    # AI
...