# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hearth — family AI workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemStructuredData } from './itemStructuredData';
import type { ItemType } from './itemType';

export interface Item {
  id: string;
  userId: string;
  rawText: string;
  type: ItemType;
  title: string;
  structuredData: ItemStructuredData;
  /** @nullable */
  total: number | null;
  createdAt: Date;
  updatedAt: Date;
}