# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hearth — family AI workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemUpdateStructuredData } from './itemUpdateStructuredData';

export interface ItemUpdate {
  title?: string;
  rawText?: string;
  structuredData?: ItemUpdateStructuredData;
}