# lib/api-zod/src/generated/types/linkInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hearth — family AI workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface LinkInput {
  url: string;
  label?: string;
}