# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hearth — family AI workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: string;
  userId: string;
  title: string;
  content: string;
  notionSynced: boolean;
  /** @nullable */
  notionPageId?: string | null;
  createdAt: Date;
  updatedAt: Date;
}